ref: code style and clean up
