+++ src/Zezo.Core.GrainInterfaces/StepGrainData.cs
@@ -8,6 +8,7 @@ namespace Zezo.Core.GrainInterfaces
     {
         public StepStatus Status { get; set; } = StepStatus.Uninitialized;
         public string Type { get; set; }
+        public string Id { get; set; }
         public Guid? ParentNode { get; set; }
         public Guid Entity {get; set; }
         public List<Guid> ChildNodes { get; set; } = new List<Guid>();
