+++ src/Zezo.Core.Grains/IContainer.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Zezo.Core.Configuration.Steps;
 using Zezo.Core.GrainInterfaces;
@@ -26,6 +27,11 @@ namespace Zezo.Core.Grains {
         void CompleteSelf(bool success);
         void MarkSelfStarted();
 
-        void SpawnChild(StepNode childConfig);
+        /// <summary>
+        /// Spawns a new StepGrain in the EntityGrain.
+        /// </summary>
+        /// <param name="childConfig">The configuration for the new Step.</param>
+        /// <returns>The Guid of the newly created StepGrain.</returns>
+        Task<Guid> SpawnStep(StepNode childConfig);
     }
 }
